Travis integration (#11)
* travis test

* Update .travis.yml

* Added test scene

* Fix command string for blender

* Moved scripts into 'addons' subfolder

Allows easier development without symbolics links which are inconvenient on windows.

* Run tests on travis

* Removed debug output

* Fixed paths and removed symlink from run_tests.py

Was not working well on windows and is not necessary anymore due to new directory structure.

* Moved travis tests to separate directory.

* Test output: Better paths on windows
