+++ .travis.yml
@@ -1,0 +1,18 @@
+language: python
+python:
+  "3.5"
+
+install:
+  - sudo add-apt-repository -y ppa:thomas-schiex/blender
+  - sudo apt-get update
+  - sudo apt-get install -y blender
+  - blender --version
+
+script:
+  python test/travis/runtest.py
+
+deploy:
+  provider: pages
+  skip_cleanup: true
+  github_token: $GITHUB_TOKEN 
+  local_dir: ouput
