--- io_scene_gltf/animation.py
@@ -1,8 +1,0 @@
-import bpy
-
-def create_action(op, idx):
-    anim = op.gltf['animations'][idx]
-    name = anim.get('name', 'animations[%d]' % idx)
-
-    action = bpy.data.actions.new(name)
-    return action
